+++ assets/src/ba_data/python/ba/_app.py
@@ -121,6 +121,7 @@ class App:
             'sr': 'Serbian',
             'uk': 'Ukrainian',
             'vi': 'Vietnamese',
+            'vec': 'Venetian',
             'hi': 'Hindi'
         }